feat: Update Options.test.jsx to use async/await for fetching scoop images
This commit updates the Options.test.jsx file to use async/await when fetching scoop images from the server. By using async/await, we ensure that the test waits for the images to be loaded before making assertions. This improves the reliability of the test and ensures that the correct number of images are displayed for each scoop option.